Changed Price from BigDecimal to Double to improve the Test functionality of the code

+++ src/main/java/com/qa/ims/persistence/dao/ItemDaoMysql.java
@@ -1,6 +1,5 @@
 package com.qa.ims.persistence.dao;
 
-import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -38,7 +37,7 @@ public class ItemDaoMysql implements Dao<Item> {
 		Item ItemFromResultSet(ResultSet resultSet) throws SQLException {
 			Long id = resultSet.getLong("id");
 			String IName = resultSet.getString("Item_name");
-			BigDecimal price = resultSet.getBigDecimal("price");
+			Double price = resultSet.getDouble("price");
 			return new Item(id, IName, price);
 		}
